--- dns/transport/local/local_resolv_stub.go
@@ -1,15 +1,0 @@
-//go:build !darwin
-
-package local
-
-import (
-	"context"
-	"os"
-
-	"github.com/sagernet/sing-box/adapter"
-	"github.com/sagernet/sing-box/log"
-)
-
-func NewResolvTransport(ctx context.Context, logger log.ContextLogger, tag string) (adapter.DNSTransport, error) {
-	return nil, os.ErrInvalid
-}
