+++ dns/transport/local/resolv_unix.go
@@ -1,4 +1,4 @@
-//go:build !windows && !darwin
+//go:build !windows
 
 package local
